fix(web): keep capability guards out of workflow node:crypto graph
Extract JobProviderActionId into a workflow-safe module so
translation-tools can import isJobCreateAllowed without pulling
provider adapters that use node:crypto into the withWorkflow bundle.
