+++ apps/hyperlocalise-web/src/lib/providers/job-provider-actions.ts
@@ -6,14 +6,10 @@ import {
   type TmsProviderCapabilityAction,
 } from "./tms-capabilities";
 import type { ExternalTmsProviderKind } from "./organization-external-tms-provider-credentials";
+import type { JobProviderActionId } from "./job-provider-action-ids";
 
-export type JobProviderActionId =
-  | "translate_with_agent"
-  | "review_with_agent"
-  | "run_qa_checks"
-  | "fix_qa_issues"
-  | "leave_provider_comment"
-  | "push_approved_changes";
+export type { JobProviderActionId } from "./job-provider-action-ids";
+export { isJobProviderActionId } from "./job-provider-action-ids";
 
 export type JobProviderActionDefinition = {
   id: JobProviderActionId;
@@ -144,14 +140,6 @@ export function getJobProviderActionAvailability(providerKind: string) {
   return jobProviderActionDefinitions.map((action) =>
     resolveActionAvailability(providerKind, action),
   );
-}
-
-const jobProviderActionIds = new Set(
-  jobProviderActionDefinitions.map((definition) => definition.id),
-);
-
-export function isJobProviderActionId(value: string): value is JobProviderActionId {
-  return jobProviderActionIds.has(value as JobProviderActionId);
 }
 
 export function getJobProviderActionDefinition(actionId: JobProviderActionId) {
